fixed bulk upload of tracks to also handle exceptions properly.
